requirements/requirements.txt
apache-beam[gcp]==2.54.0
requests==2.31.0
google-cloud-storage==2.16.0
google-cloud-bigquery==3.17.2